ia-engine/pyproject.toml
[tool.poetry]
name = "email-studio-ia-engine"
version = "0.1.0"
description = "Motor de IA para Email Studio (FastAPI + Python)."
authors = ["Equipo Martech Banco BICE"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"

fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}

pydantic = "^2.8.0"

anthropic = "^0.40.0"
openai = "^1.40.0"

pandas = "^2.2.0"
numpy = "^2.0.0"
tabulate = "^0.9.0"

python-dotenv = "^1.0.1"

google-auth = "^2.32.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
